Like I love when this happens. I write code that works but doesnt pass the checker. And the checker wont tell me why

+++ python_async_function/0-basic_async_syntax.py
@@ -5,7 +5,7 @@ import random
 from typing import *
 
 
-async def wait_random(max_delay: float = 10) -> float :
+async def wait_random(max_delay: int = 10) -> float :
     """Uses async to find to delay"""
     delay = random.uniform(0, max_delay)
     await asyncio.sleep(delay)
